+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **DOCS**: Fix missing `MainAxisSize.min` in `ShadCard` notification example.
 - **DOCS**: Adjust sheet docs example, with some spacing between buttons.
 - **FIX**: Fix the hit test behavior of `ShadContextMenuRegion`.
+- **FEAT**: Expose `hitTestBehavior` from `ShadContextMenuRegion`, defaults to `HitTestBehavior.opaque` (thanks to @NonymousMorlock).
 
 ## 0.27.4